Renamed ObjectRefSelectionDialog to SelectionDialog
