+++ Source/Editor/DualityEditor/DualityEditor.csproj
@@ -121,11 +121,11 @@
     <Compile Include="Extensibility\Previews\PreviewQuery.cs" />
     <Compile Include="Extensibility\Previews\IPreviewQuery.cs" />
     <Compile Include="Extensibility\Previews\PreviewSizeMode.cs" />
-    <Compile Include="Forms\ObjectRefSelectionDialog.cs">
+    <Compile Include="Forms\SelectionDialog.cs">
       <SubType>Form</SubType>
     </Compile>
-    <Compile Include="Forms\ObjectRefSelectionDialog.Designer.cs">
-      <DependentUpon>ObjectRefSelectionDialog.cs</DependentUpon>
+    <Compile Include="Forms\SelectionDialog.Designer.cs">
+      <DependentUpon>SelectionDialog.cs</DependentUpon>
     </Compile>
     <Compile Include="Forms\CreateObjectDialog.cs">
       <SubType>Form</SubType>
@@ -329,8 +329,8 @@
     <EmbeddedResource Include="AssetManagement\SelectAssetImporterDialog.resx">
       <DependentUpon>SelectAssetImporterDialog.cs</DependentUpon>
     </EmbeddedResource>
-    <EmbeddedResource Include="Forms\ObjectRefSelectionDialog.resx">
-      <DependentUpon>ObjectRefSelectionDialog.cs</DependentUpon>
+    <EmbeddedResource Include="Forms\SelectionDialog.resx">
+      <DependentUpon>SelectionDialog.cs</DependentUpon>
     </EmbeddedResource>
     <EmbeddedResource Include="Forms\CreateObjectDialog.resx">
       <DependentUpon>CreateObjectDialog.cs</DependentUpon>
