+++ Source/Plugins/EditorBase/PropertyEditors/AddComponentPropertyEditor.cs
@@ -89,7 +89,7 @@ namespace Duality.Editor.Plugins.Base.PropertyEditors
 		}
 		private void OnAddComponentClicked()
 		{
-			ObjectRefSelectionDialog compTypeSelector = new ObjectRefSelectionDialog
+			SelectionDialog compTypeSelector = new SelectionDialog
 			{
 				FilteredType = typeof(Component),
 				SelectType = true
