+++ Source/Plugins/EditorBase/PropertyEditors/ObjectRefPropertyEditor.cs
@@ -393,7 +393,7 @@ namespace Duality.Editor.Plugins.Base.PropertyEditors
 
 		protected void ShowContentSelectionDialog()
 		{
-			ObjectRefSelectionDialog resourceSelectionForm = new ObjectRefSelectionDialog
+			SelectionDialog resourceSelectionForm = new SelectionDialog
 			{
 				FilteredType = this.FilteredType,
 				ResourcePath = this.ReferenceName
